requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
Django==4.0.3
django-cors-headers==3.11.0
django-cors-middleware==1.5.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
entrypoints==0.4
executing==0.8.3
fonttools==4.31.2
ipykernel==6.9.2
ipython==8.1.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.4.0
lightgbm==3.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.22.3
packaging==21.3
pandas==1.4.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
psutil==5.9.0
pure-eval==0.2.2
Pygments==2.11.2
PyJWT==2.3.0
pylab-sdk==1.3.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
pyzmq==22.3.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
stack-data==0.2.0
threadpoolctl==3.1.0
tornado==6.1
traitlets==5.1.1
tzdata==2022.1
wcwidth==0.2.5
whitenoise==6.0.0
